State when asking question about probe-rs debugging.
probe-rs runs from the command line as a runner (cargo run), and sends defmt info to the console.
I fiddled with launch.json to get debugging partially working, but it crashes when I continue or step the code.  Asked a question on the probe-rs GitHub discussion board about it 2024-10-12.
Changed main blink rate to ensure that the VSCode extension was flashing the board, and changed defmt macros from info to warn because the launch said that it would be printing warn.  There's probably a configuration somewhere to choose that.